Added ticket save validation
Removed old code to update Impersonation Email since we deprecated this field on the user profile. Added validation so that ticket can't be set onto pending unless required fields are filled and ticket can't be set to on hold unless there's a date in the future in the reminder field.